\name{conc}
\alias{conc}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{A child function to calcuate concordance
}
\description{
This is a child function (R-wrapper) called in the main functions in this package. 
}
\usage{
conc(X, D, W, R)
}
\arguments{
  \item{X}{time
}
  \item{D}{event indicator (1:event, 0:censor)
}
  \item{W}{weight
}
  \item{R}{risk score
}
}