\name{unoU2P}
\alias{unoU2P}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{A child function used in Inf.Cval and Inf.Cval.Delta
}
\description{
This is a child function (R-wrapper) used in the main functions of this package. 
}
\usage{
unoU2P(A, B)
}
\arguments{
  \item{A}{A nxp matrix
}
  \item{B}{A nx1 vector
}
}