\name{VEC2MAT}
\alias{VEC2MAT}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Vector to Matrix
}
\description{
Making a (PxM) matrix (x',x',x',...,x') from a (Px1) vector, x
}
\usage{
VEC2MAT(vc, dm)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{vc}{A vector 
}
  \item{dm}{The number of row of the matrix returned 
}
}
