\name{unoU2P}
\alias{unoU2P}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{A child function used in Inf.Cval and Inf.Cval.Delta
}
\description{
This is a child function (R-wrapper) used in the main functions of this package. 
}
\usage{
unoU2P(A, B)
}
\arguments{
  \item{A}{A nxp matrix
}
  \item{B}{A nx1 vector
}
}
\details{
%%This function returns (sum(A[i,]+A[j,])B[i]B[j])/2/(n choose 2).
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
